requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-astradb
google-generativeai
faiss-cpu
bandit

# Testing
pytest
